Making a start on the work for #271 on the main spec.
